Cleanup of code and better organisation into propagate_segment.c and mcfret4.c #136
