+++ src/NISE_subs.c
@@ -10,6 +10,7 @@
 #include "read_trajectory.h"
 #include "randomlib.h"
 #include "util/asprintf.h"
+#include <cblas.h>
 
 // Subroutines for nonadiabatic code
 
@@ -55,6 +56,13 @@ void clearvec_double(double* a, int N) {
     int i;
     for (i = 0; i < N; i++) a[i] = 0;
 }
+
+// Set all elements of a vector to zero, here for integers
+void clearvec_int(int *a, int N) {
+    int i;
+    for (i = 0; i < N; i++) a[i] = 0;
+}
+
 
 // Construct unit matrix
 void unitmat(float *a,int N){
@@ -119,6 +127,113 @@ void trans_matrix_on_vector(float *c,float *vr,float *vi,int N){
     copyvec(xi,vi,N);
     free(xr);
     free(xi);
+}
+/* Find the trace of a matrix product */
+float matrix_mul_traced_DA(float *A, float *B, int N_i, int N_i3){
+    /* Calculate the trace of the product of two differently shaped matrices of dimension N_i * Ni_3 */
+    /* Directly computing the trace greatly reduces the number of operations needed */
+    /* from N^3 to N^2 */
+    // N_i3 is the shared dimension of the matrices A & B
+    // N_i is the dimension of the resulting square matrix
+    int i,i3;
+    float the_trace;
+
+    the_trace = 0;
+
+    for (i=0;i<N_i;i++){
+   	    for (i3=0;i3<N_i3;i3++){
+            the_trace += A[i*N_i3+i3] * B[i3*N_i+i];
+	    }
+    }
+    return the_trace;
+}
+
+// void complex_matrix_product(float *A_re, float *A_im, float *B_re, float *B_im, float *C_re, float *C_im,int N_1,int N_2,int N_3){
+
+//     int i1, i2, i3;
+//     float Aim_i1i3, Are_i1i3;
+//     clearvec(C_re,N_1*N_2);
+//     clearvec(C_im,N_1*N_2);
+
+// // can make paralllel but check for loop order: warnings recieved
+// #pragma omp parallel for private(i2,i3,Aim_i1i3,Are_i1i3)
+// for (i1=0;i1<N_1;i1++){
+//     for (i3=0;i3<N_3;i3++){
+// 	    Aim_i1i3 = A_im[i1*N_3+i3];
+// 	    Are_i1i3 = A_re[i1*N_3+i3];
+//             for (i2=0;i2<N_2;i2++){
+//                     C_re[i1*N_2+i2] += Are_i1i3 * B_re[i3*N_2+i2] - Aim_i1i3 * B_im[i3*N_2+i2];
+//                     C_im[i1*N_2+i2] += Are_i1i3 * B_im[i3*N_2+i2] + Aim_i1i3 * B_re[i3*N_2+i2];
+//                 }
+//             }
+//         }
+// }
+
+/* Computes the complex matrix product of two matrices, */
+/* After converting to complex arrays */
+/* Only recommended for large matrices (i.e. N >1000) */
+void complex_matrix_product(float *A_re, float *A_im,
+                            float *B_re, float *B_im,
+                            float *C_re, float *C_im,
+                            int N_1, int N_2, int N_3)
+{
+    // Row-major matrices:
+    // A: N_1 x N_3
+    // B: N_3 x N_2
+    // C: N_1 x N_2
+
+    const int sizeA = N_1 * N_3;
+    const int sizeB = N_3 * N_2;
+    const int sizeC = N_1 * N_2;
+
+    // Allocate interleaved complex buffers
+    float *A = (float*) malloc(sizeof(float) * 2 * sizeA);
+    float *B = (float*) malloc(sizeof(float) * 2 * sizeB);
+    float *C = (float*) malloc(sizeof(float) * 2 * sizeC);
+
+    if (!A || !B || !C) {
+        free(A);
+        free(B);
+        free(C);
+        return; // allocation failed
+    }
+
+    // Convert A to interleaved complex
+    for (int i = 0; i < sizeA; i++) {
+        A[2*i]     = A_re[i];
+        A[2*i + 1] = A_im[i];
+    }
+
+    // Convert B
+    for (int i = 0; i < sizeB; i++) {
+        B[2*i]     = B_re[i];
+        B[2*i + 1] = B_im[i];
+    }
+
+    // Zero C
+    memset(C, 0, sizeof(float) * 2 * sizeC);
+
+    const float alpha[2] = {1.0f, 0.0f};
+    const float beta[2]  = {0.0f, 0.0f};
+
+    cblas_cgemm(CblasRowMajor,
+                CblasNoTrans, CblasNoTrans,
+                N_1, N_2, N_3,
+                alpha,
+                A, N_3,
+                B, N_2,
+                beta,
+                C, N_2);
+
+    // Convert result back to split format
+    for (int i = 0; i < sizeC; i++) {
+        C_re[i] = C[2*i];
+        C_im[i] = C[2*i + 1];
+    }
+
+    free(A);
+    free(B);
+    free(C);
 }
 
 /* Find the norm for a complex vector */
