+++ src/NISE_subs.h
@@ -8,10 +8,13 @@ void not_parallel();
 void copyvec(float *a,float *b,int N);
 void clearvec(float *a,int N);
 void clearvec_double(double* a, int N);
+void clearvec_int(int *a, int N);
 void unitmat(float *a,int N);
 void vector_on_vector(float *rr,float *ir,float *vr,float *vi,int N);
 void matrix_on_vector(float *c,float *vr,float *vi,int N);
 void trans_matrix_on_vector(float *c,float *vr,float *vi,int N);
+float matrix_mul_traced_DA(float *A, float *B, int N_i, int N_i3);
+void complex_matrix_product(float *A_re, float *A_im, float *B_re, float *B_im, float *C_re, float *C_im,int N1,int N2,int N3);
 float find_norm(float *phi_r,float *phi_i,int N);
 void re_normalize(float *phi_r,float *phi_i,int N,float norm);
 void log_item(char* msgFormat, ...);
